rename files to use dash to keep package conventions
