+++ mobile/packages/common/index.ts
@@ -20,7 +20,7 @@ export * from './fixtures/SuspenseBoundary'
 export * from './utils/arrays'
 export * from './utils/hex'
 export * from './utils/monads'
-export * from './utils/parseNumberFromText'
+export * from './utils/parse-number-from-text'
 export * from './utils/parsers'
 export * from './utils/predicates'
 export * from './utils/promises'
