+++ mobile/packages/common/utils/parse-number-from-text.ts
